requirements.txt
Flask==2.0.2
python-dotenv==0.19.2
gunicorn==20.1.0
